vcc: Run compile time recursion check also on dynamic-only SUBs
We previously missed to run the static recursion check on SUBs which
are only referenced symbolically (and thus might be called
dynamically).

This changes the error reporting tested in v00021.vtc because the
recursion is detected on the recursing subs alone before the call tree
from vcl_recv {} is walked.